Reduce coupling between modules
For more convenient testing it is necessary to reduce coupling between modules so they could be tested on their own.
