+++ Validator.js
@@ -1,8 +1,6 @@
 'use strict';
 
-const fs = require('fs').promises;
-const path = require('path');
-const { regexps, tagsHtml, inputExtensions, outputExtensions } = require('./config/constants');
+const { regexps, tagsHtml } = require('./config/constants');
 
 class Validator {
   constructor () {
@@ -10,32 +8,6 @@ class Validator {
     this.inPreformattedText = false;
     this.mdTags = Object.entries(tagsHtml).map(([, mdTag]) => mdTag.md);
     this.isClosedRegexps = Object.entries(regexps).map(([, regexp]) => regexp);
-  }
-
-  async validateMdFilepath (mdFilePath) {
-    await this.checkAccess(mdFilePath);
-    this.checkExtension(mdFilePath, true);
-  }
-
-  validateOutputPath (filePath) {
-    this.checkExtension(filePath, false);
-  }
-
-  async checkAccess (mdFilePath) {
-    try {
-      await fs.access(mdFilePath);
-    } catch (err) {
-      throw new Error(`Error: No access to "${mdFilePath}". Wrong option's format or non-existing file.`);
-    }
-  }
-
-  checkExtension (filePath, isInput) {
-    const ext = path.extname(filePath);
-    const valid = isInput ? inputExtensions.includes(ext) : outputExtensions.includes(ext);
-    const allowedExtensions = isInput ? inputExtensions : outputExtensions;
-    if (!valid) {
-      throw new Error(`Error: The file "${filePath}" has wrong extension. Only ${allowedExtensions} allowed`);
-    }
   }
 
   validateMdContent (line) {
